moved FieldAdapter to field_models.py
